client/requirements.txt
bidict==0.21.2
certifi==2021.5.30
charset-normalizer==2.0.3
coloredlogs==15.0.1
humanfriendly==9.2
idna==3.2
pyreadline==2.1
python-engineio==4.2.0
python-socketio==5.3.0
requests==2.26.0
urllib3==1.26.6
websocket-client==1.1.0
